move interactive test to test/plotting/plotting.py
