+++ tests/plotting/test_plotting.py
@@ -1370,3 +1370,16 @@ def test_index_vs_loc():
         assert pl.index_to_loc(np.int_(val)) == val
         assert pl.loc_to_index(val) == val
         assert pl.loc_to_index(np.int_(val)) == val
+
+
+@skip_no_plotting
+def test_interactive_update():
+    # Regression test for #1053
+    p = pyvista.Plotter()
+    p.show(interactive_update=True)
+    assert isinstance(p.iren, vtk.vtkRenderWindowInteractor)
+    p.close()
+
+    p = pyvista.Plotter()
+    with pytest.warns(UserWarning):
+        p.show(auto_close=True, interactive_update=True)
